+++ src/ModelPhysics/Energy/energy_types.jl
@@ -1,4 +1,6 @@
-export AbstractEnergyModel, Energy
+# export AbstractEnergyModel, Energy
+export Energy
+export AbstractEnergyModel
 export Isothermal
 
 abstract type AbstractEnergyModel end
@@ -11,15 +13,21 @@ end
 
 # Isothermal
 
-struct Isothermal <: AbstractEnergyModel end 
+struct Isothermal{T<:AbstractScalarField} <: AbstractEnergyModel 
+    T::T
+end 
 Adapt.Adapt.@adapt_structure Isothermal
 
-Energy{Isothermal}() = begin
-    args = nothing
-    ARGS = typeof(args)
-    Energy{Isothermal,ARGS}(args)
+# Set the defualt value unless user specifies it - compatible with other solvers and energy models.
+Energy{Isothermal}(; T=300.0) = begin
+    coeffs = (T=T,)
+    ARG = typeof(coeffs)
+    Energy{Isothermal,ARG}(coeffs)
 end
 
 (energy::Energy{EnergyModel, ARG})(mesh, fluid) where {EnergyModel<:Isothermal,ARG} = begin
-    nothing
+    T_val = energy.args.T
+    
+    T = ConstantScalar(T_val)
+    Isothermal(T)
 end